before switch index to u32

+++ src/models/ReservationProdiverData.ts
@@ -3,7 +3,7 @@ import { u128 } from '@btc-vision/as-bignum/assembly';
 
 export class ReservationProviderData {
     constructor(
-        public readonly providerIndex: u64,
+        public readonly providerIndex: u32,
         public readonly providedAmount: u128,
         public readonly providerType: ProviderTypes,
     ) {}
